Add Company page component
